OXDEV-2642 - renaming
